move search to app router
