+++ samples/next-with-stripe/app/layout.tsx
@@ -1,5 +1,9 @@
 import { WithStanzaContextName } from '@getstanza/react'
-import WithStanza from './withStanza'
+import WithStanza from '../components/WithStanza'
+import StripeCartProvider from '../components/StripeCartProvider'
+import Layout from '../components/Layout'
+
+import '../styles.css'
 
 export default function RootLayout ({
   children
@@ -8,13 +12,17 @@ export default function RootLayout ({
 }) {
   return (
     <html lang="en">
-    <body>
-    <WithStanza>
-      <WithStanzaContextName name="main">
-        { children }
-      </WithStanzaContextName>
-    </WithStanza>
-    </body>
+      <body>
+        <WithStanza>
+          <WithStanzaContextName name="main">
+            <StripeCartProvider>
+              <Layout title="Stanza Toy Store">
+                <div className="page-container">{children}</div>
+              </Layout>
+            </StripeCartProvider>
+          </WithStanzaContextName>
+        </WithStanza>
+      </body>
     </html>
   )
 }
