+++ samples/next-with-stripe/app/search/page.tsx
@@ -1,0 +1,7 @@
+import SearchPageComponent from './search'
+
+const SearchPage = () => {
+  return <SearchPageComponent/>
+}
+
+export default SearchPage
