+++ samples/next-with-stripe/components/Layout.tsx
@@ -1,3 +1,4 @@
+'use client'
 import React, { type ReactNode } from 'react'
 import Head from 'next/head'
 import Link from 'next/link'
