+++ samples/next-with-stripe/components/StripeCartProvider.tsx
@@ -1,3 +1,4 @@
+'use client'
 import React, { type ReactNode } from 'react'
 import { CartProvider } from 'use-shopping-cart'
 import * as config from '../config'
