+++ samples/next-with-stripe/components/WithStanza.tsx
