+++ samples/next-with-stripe/pages/_app.tsx
@@ -1,43 +1,25 @@
-import {
-  createStanzaInstance,
-  StanzaProvider,
-  WithStanzaContextName
-} from '@getstanza/react'
+import { WithStanzaContextName } from '@getstanza/react'
 import { type AppProps } from 'next/app'
 import React from 'react'
 import Layout from '../components/Layout'
 import StripeCartProvider from '../components/StripeCartProvider'
-import { config } from '../stanzaConfig'
 
 import '../styles.css'
-import Link from 'next/link'
-
-let loadPromise: Promise<any> = Promise.resolve()
-if (process.env.NODE_ENV === 'development') {
-  // const mswMock = import('../msw/mock')
-  // loadPromise = mswMock.then(async module => module.initMocks())
-  loadPromise = Promise.resolve()
-}
-
-const stanzaInstance = createStanzaInstance({
-  ...config,
-  pollDelay: loadPromise
-})
+import WithStanza from '../components/WithStanza'
 
 function MyApp ({ Component, pageProps }: AppProps) {
   return (
-    <StanzaProvider instance={stanzaInstance}>
+    <WithStanza>
       <WithStanzaContextName name="main">
         <StripeCartProvider>
           <Layout title="Stanza Toy Store">
             <div className="page-container">
-              <Link href={'/test'}>Test Page</Link>
               <Component {...pageProps} />
             </div>
           </Layout>
         </StripeCartProvider>
       </WithStanzaContextName>
-    </StanzaProvider>
+    </WithStanza>
   )
 }
