+++ samples/next-with-stripe/styles.css
@@ -29,7 +29,7 @@ body {
   -webkit-tap-highlight-color: rgba(0, 0, 0, 0);
 }
 
-#__next {
+body {
   display: flex;
   justify-content: center;
 }
